card downloader as its own module
